Added website feature documentation for @ToString and @EqualsAndHashCode, and modified the docs for @Data to refer to the docs of these new annotations.
The build script for the website has been updated to clean out the website build dir first, so files that used to exist but have now been tossed are no longer there.

There's also a special website-no-videos target which builds a website deployable without the videos. This makes the upload a lot faster if the videos haven't changed.